make components and have pages use them and redo description
